requirements.txt
numpy==1.18.0
opencv-python==4.1.2.30
Pillow==6.2.1
pynput==1.6.0
pyobjc-core==6.1
pyobjc-framework-Cocoa==6.1
pyobjc-framework-Quartz==6.1
six==1.13.0